HPA-17: add tests for the supplier variants of runTransacted + fixes
Made the exceptionHandler of the supplier variant of `runTransacted` be of type `Function` instead of `Consumer`, to give the exception handler the ability to recover and send a proper return value, based on the exception.
While this exception handler is run for nested transactions, only the exception handler for the outer-most call to `runTransacted` can determine the actual return value.
The inner exception handlers technically could give extra clues to the outer handler through the exception itself.