HPA-17: move locking to the runTransacted method to prevent exceptionHandlerStack corruption and make sure that the lock is released when an exception is thrown
The release lock is reused, but this is fine since this is the same ReentrantLock under the hood. The only minor inconvenience is that if you somehow manage to release one too many times, that you'll get an IllegalMonitoringStateException, but since the LockManager is package private, this is near impossible to occur, as long as every acquireLock/tryAcquiringLock is matched up with a lock.release in JpaNestableTransaction.