HPA-17: change queries with the IN keyword so that they are performed in batches and add getMutableResultList to the QueryService
OpenJPA was having issues with queries with the IN keyword if there were more than a couple of thousand parameters. These queries have now been cut up into chunks and are combined afterwards. The size of these chunks is made configurable

getMutableResultList was added because by default it is not possible to modify the list returned by query.getResultList. This functionality was needed, so this method was provided.